moved chat and game to specific component folders in component folder
