+++ frontend/src/App.tsx
@@ -3,24 +3,21 @@ import {
   BrowserRouter as Router,
   Routes,
   Route,
-  Navigate,
 } from "react-router-dom";
 import Login from "./components/Login";
 import UserPage from "./components/UserPage";
 import AuthCallback from "./components/AuthCallback";
 import ProtectedRoute from "./components/ProtectedRoute";
-import Chat from "./chat/Chat";
-import App_game from "./App_game";
+import Chat from "./components/chat/Chat";
+import GameContainer from "./components/game/GameContainer";
 
-function App(): JSX.Element {
+const App: React.FC = () => {
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<Navigate to="/login" />} />
+        <Route index element={<Login />} />
         <Route path="/login" element={<Login />} />
         <Route path="/auth/42/callback" element={<AuthCallback />} />
-        {/* <Route path="/game" element={<App_game />} /> */}
-        {/* <Route path="/chat" element={<Chat />} /> */}
         <Route
           path="/userpage"
           element={
@@ -41,13 +38,13 @@ function App(): JSX.Element {
           path="/game"
           element={
             <ProtectedRoute>
-              <App_game />
+              <GameContainer />
             </ProtectedRoute>
           }
         />
       </Routes>
     </Router>
   );
-}
+};
 
 export default App;
