+++ frontend/src/components/chat/mountain-valley.jpg
