+++ frontend/src/components/chat/talking_cat_ab.jpeg
