+++ frontend/src/components/game/Lobby.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { GameMode, LobbyProps } from "../types";
+import { GameMode, LobbyProps } from "../../types";
 import CustomGameCreation from "./CustomGameCreation";
 import { useNavigate } from "react-router-dom";
 import {
@@ -11,7 +11,7 @@ import {
   offMatchFound,
   onQueueStatus,
   offQueueStatus,
-} from "../socket";
+} from "../../socket";
 
 const Lobby: React.FC<LobbyProps> = ({
   onGameStart,
