+++ frontend/src/components/game/Scoreboard.tsx
