+++ frontend/src/global.interface.tsx
@@ -1,14 +1,14 @@
 export interface User {
-	userId: number;
-	username: string;
-	email: string;
-	oauthToken: string;
-	oauthRefreshToken: string;
-	provider: string;
-	displayName: string;
-	avatar?: string; // Add avatar field to the User model
-	  imageType: string;
-	  imageName?: string;
-	  imageData?: ArrayBuffer;
-	  imageString?: string;
-  }
+  userId: number;
+  username: string;
+  email: string;
+  oauthToken: string;
+  oauthRefreshToken: string;
+  provider: string;
+  displayName: string;
+  avatar?: string; // Add avatar field to the User model
+  imageType: string;
+  imageName?: string;
+  imageData?: ArrayBuffer;
+  imageString?: string;
+}
